os-snap: code cleanup
