+++ os-snap/test/snap_service_test.dart
@@ -6,7 +6,6 @@ import 'package:redstone/mocks.dart';
 import 'package:mongo_dart/mongo_dart.dart';
 import 'package:os_common/os_common.dart';
 import 'package:os_common/os_common_server.dart';
-import 'package:os_snap/os_snap.dart';
 
 const uri = 'mongodb://localhost/test';
 Db _db = new Db(uri);
@@ -20,19 +19,19 @@ Snap _snap1 = new Snap(_pauline, [_seb], 'data:image/jpg;base64,/9j/4AAQSkZJR', 
 Snap _snap2 = new Snap(_baptiste, [_pauline, _seb], 'data:image/jpg;base64,/9j/4AAQSkZJR', 10, '563b80c33786c930f70e4001');
 
 main() {
-  
+
   setUp(() {
     app.setupConsoleLog();
     app.addModule(new Module()..bind(Db, toValue: _db));
     app.addPlugin(ObjectMapper);
     app.setUp([#os_common, #os_snap]);
-    
+
     return _db.open().then((_) => _snaps.drop().then((_) => _snaps.insertAll([_snap1.toJson(), _snap2.toJson()])));
   });
 
   tearDown(() {
     app.tearDown();
-    return _db.close(); 
+    return _db.close();
   });
 
   test('Create', () {
@@ -44,9 +43,9 @@ main() {
       return _snaps.findOne(where.eq('_id', snap.id)).then((_) => expect(_, isNotNull));
     });
   });
-  
+
   test('Get by id', () {
-    var req = new MockRequest('/snap/${_snap1.id}');  
+    var req = new MockRequest('/snap/${_snap1.id}');
     return app.dispatch(req).then((resp) {
       expect(resp.statusCode, equals(HttpStatus.OK));
       Snap snap = new Snap.fromJson(resp.mockContent);
@@ -54,9 +53,9 @@ main() {
       expect(snap.id, equals(_snap1.id));
     });
   });
-  
+
   test('Get by ids', () {
-    var req = new MockRequest('/snap/${_snap1.id},${_snap2.id}');  
+    var req = new MockRequest('/snap/${_snap1.id},${_snap2.id}');
     return app.dispatch(req).then((resp) {
       expect(resp.statusCode, equals(HttpStatus.OK));
       List<Snap> snaps = Snap.fromJsonList(resp.mockContent);
@@ -64,9 +63,9 @@ main() {
       expect(snaps.length, equals(2));
     });
   });
-  
+
   test('Get by author', () {
-    var req = new MockRequest('/snap/sent/${_pauline.id}');  
+    var req = new MockRequest('/snap/sent/${_pauline.id}');
     return app.dispatch(req).then((resp) {
       expect(resp.statusCode, equals(HttpStatus.OK));
       List<Snap> snaps = Snap.fromJsonList(resp.mockContent);
@@ -75,9 +74,9 @@ main() {
       expect(snaps[0].id, equals(_snap1.id));
     });
   });
-  
+
   test('Get by recipient', () {
-    var req = new MockRequest('/snap/received/${_seb.id}');  
+    var req = new MockRequest('/snap/received/${_seb.id}');
     return app.dispatch(req).then((resp) {
       expect(resp.statusCode, equals(HttpStatus.OK));
       List<Snap> snaps = Snap.fromJsonList(resp.mockContent);
@@ -85,9 +84,9 @@ main() {
       expect(snaps.length, equals(2));
     });
   });
-  
+
   test('Delete by id', () {
-    var req = new MockRequest('/snap/${_snap1.id}', method: app.DELETE);  
+    var req = new MockRequest('/snap/${_snap1.id}', method: app.DELETE);
     return app.dispatch(req).then((resp) {
       expect(resp.statusCode, equals(HttpStatus.NO_CONTENT));
       return _snaps.findOne(where.eq('_id', _snap1.id)).then((_) => expect(_, isNull));
